Removed an int cast on a WIN32 DWORD. The cast was seemingly redundant per clang-tidy? But DWORD!=int...

+++ reproc/src/reproc.c
@@ -663,7 +663,7 @@ int reproc_pid(reproc_t *process)
   // However process IDs are never negative and its
   // extremely unlikely a processID on windows would be
   // larger than 200k.
-  return (int) GetProcessId(process->handle);
+  return GetProcessId(process->handle);
 #else
   // On posix return the handle directly as this is
   // the process ID.
